Branching, merging, and conflicts
This adds a new chapter, dealing with branching, merging, and conflicts.

Originally I had templates in this chapter, but I think it's too long of a tangent, so I've also moved that content out to another chapter on customization. We'll see if it stays there or moves around more.

Also, I wanted to take the conflict out of "The edit workflow" chapter. I don't think it's fair to show one workflow with conflicts and one without conflicts, as it makes that workflow look worse. Also, it means I can use that conflicting examples as the foundation of this chapter.