Add missing value '@' for jj new -B
It seems that jj new -B / --insert-before has no default and a value needs
to be provided explicitly. Running jj new -B -m ... produces an error message:

error: a value is required for '--insert-before <INSERT_BEFORE>' but none was supplied

Tested on jj 0.22.0. This is probably linked to changes introduced in 0.18.0:
https://github.com/martinvonz/jj/blob/main/CHANGELOG.md#0180---2024-06-05