Use more accurate wording of what's outdated in "Rebasing a PR"
When GitHub marks a comment thread as "Outdated" it means the diff being
commented on has changed. If you comment on a line and then the author
rebases without changing that line, GitHub will _not_ mark the comment
as outdated (even though the original commit you commented on is
gone/outdated).