Update Remotes page title
"Working with remotes, AKA, Github" implies that "remotes" are "also known as Github". However as the rest of doc states, Github is just one example of a `git` server which can be used as a Remote. 

Rewording to "Working with remotes, e.g., Github" clarifies in the title that Github is not "The Remote" but "A Remote".