Replace stub (^C^V) intro for "More advanced workflows"
This section had an intro paragraph copied and pasted from the
previous section (Sharing your code with others)--which is how I
create new sections, too, to be fair--so I wrote a simple summary.
The primary author will probably want to replace it in his own
voice, but this is a more servicible stub in the meantime.